Autosave the current buffer to a file on every change
By autosaving the current buffer to a file, emscripten
will let us hook into a file save event and process the contents
of the file via JavaScript.

I may change this later to specifically generate a giant
char* of the current buffer's memory line and send that
to the JavaScript runtime instead, but since we're using
MEMFS via emscripten this may be enough.